+++ src/onegov.user/onegov/user/forms/__init__.py
@@ -2,10 +2,12 @@ from onegov.user.forms.login import LoginForm
 from onegov.user.forms.registration import RegistrationForm
 from onegov.user.forms.reset_password import PasswordResetForm
 from onegov.user.forms.reset_password import RequestPasswordResetForm
+from onegov.user.forms.signup_link import SignupLinkForm
 
 __all__ = [
     'LoginForm',
     'PasswordResetForm',
     'RegistrationForm',
     'RequestPasswordResetForm',
+    'SignupLinkForm'
 ]