+++ src/onegov.user/onegov/user/tests/test_auth.py
@@ -1,14 +1,27 @@
+import pytest
+import time
 import transaction
 
+from datetime import datetime, timedelta
 from onegov.core import Framework
+from onegov.core.utils import Bunch
 from more.itsdangerous import IdentityPolicy
 from onegov.user import (
     Auth, is_valid_yubikey, is_valid_yubikey_format,
     UserCollection, yubikey_otp_to_serial
 )
+from onegov.user.errors import ExpiredSignupLinkError
 from webtest import TestApp as Client
 from unittest.mock import patch
 from yubico_client import Yubico
+
+
+class DummyPostData(dict):
+    def getlist(self, key):
+        v = self[key]
+        if not isinstance(v, (list, tuple)):
+            v = [v]
+        return v
 
 
 def test_auth_login(session):
@@ -206,3 +219,75 @@ def test_auth_integration(session):
     assert response.status_code == 302
     assert response.location == 'http://localhost/go'
     assert response.headers['Set-Cookie'].startswith('userid=;')
+
+
+def test_signup_token_data(session):
+    auth = Auth(session, 'foo', signup_token_secret='bar')
+    assert auth.new_signup_token('admin')
+
+    token = auth.new_signup_token('admin', max_age=1)
+    data = auth.decode_signup_token(token)
+    assert data['role'] == 'admin'
+    assert data['max_uses'] == 1
+
+    before = int((datetime.utcnow() - timedelta(seconds=1)).timestamp())
+    after = int((datetime.utcnow() + timedelta(seconds=1)).timestamp())
+    assert before <= data['expires'] <= after
+
+
+def test_signup_max_uses(session):
+    auth = Auth(session, 'foo', signup_token_secret='bar')
+    auth.signup_token = auth.new_signup_token('admin', max_age=10, max_uses=1)
+
+    foo = auth.register(
+        form=Bunch(
+            username=Bunch(data='foo@example.org'),
+            password=Bunch(data='correct horse'),
+        ),
+        request=Bunch(
+            client_addr='127.0.0.1'
+        )
+    )
+
+    assert foo.role == 'admin'
+
+    with pytest.raises(ExpiredSignupLinkError):
+        auth.register(
+            form=Bunch(
+                username=Bunch(data='bar@example.org'),
+                password=Bunch(data='battery staple'),
+            ),
+            request=Bunch(
+                client_addr='127.0.0.1'
+            )
+        )
+
+
+def test_signup_expired(session):
+    auth = Auth(session, 'foo', signup_token_secret='bar')
+    auth.signup_token = auth.new_signup_token('admin', max_age=1, max_uses=2)
+
+    foo = auth.register(
+        form=Bunch(
+            username=Bunch(data='foo@example.org'),
+            password=Bunch(data='correct horse'),
+        ),
+        request=Bunch(
+            client_addr='127.0.0.1'
+        )
+    )
+
+    assert foo.role == 'admin'
+
+    time.sleep(2)
+
+    with pytest.raises(ExpiredSignupLinkError):
+        auth.register(
+            form=Bunch(
+                username=Bunch(data='bar@example.org'),
+                password=Bunch(data='battery staple'),
+            ),
+            request=Bunch(
+                client_addr='127.0.0.1'
+            )
+        )